pyproject.toml
[tool.poetry]
name = "notifier"
version = "0.1.0"
description = ""
authors = ["Semaputnik <semyon.putnikov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.7"
websockets = "^12.0"
asyncio = "^3.4.3"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.3"
jinja2 = "^3.1.2"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
